src/tasks/imagery/requirements.txt
albumentations==1.2.0
google-cloud-storage==2.5.0
opencv-python-headless>=4.5.2.54
Flask==2.1.0; python_version > '3.6'
Flask==2.0.3; python_version < '3.7'
gunicorn==20.1.0
google-cloud-tasks==2.7.1
google-cloud-storage==2.5.0
google-cloud-bigquery==3.3.3